Move DiskIoMetrics out of IndexMetricsEntry.
