+++ searchcore/src/vespa/searchcore/proton/metrics/index_metrics_entry.h
@@ -2,9 +2,8 @@
 
 #pragma once
 
+#include "disk_io_metrics.h"
 #include "field_metrics_entry.h"
-
-namespace search { class CacheDiskIoStats; }
 
 namespace proton {
 
@@ -13,24 +12,6 @@ namespace proton {
  * disk indexes and memory indexes.
  */
 class IndexMetricsEntry : public FieldMetricsEntry {
-    class DiskIoMetrics : public metrics::MetricSet {
-        class SearchMetrics : public metrics::MetricSet {
-            metrics::LongValueMetric _read_bytes;
-            metrics::LongValueMetric _cached_read_bytes;
-        public:
-            explicit SearchMetrics(metrics::MetricSet* parent);
-            ~SearchMetrics() override;
-            void update(const search::CacheDiskIoStats& cache_disk_io_stats);
-        };
-
-        SearchMetrics _search;
-
-    public:
-        explicit DiskIoMetrics(metrics::MetricSet* parent);
-        ~DiskIoMetrics() override;
-        void update(const search::CacheDiskIoStats& cache_disk_io_stats) { _search.update(cache_disk_io_stats); }
-    };
-
     DiskIoMetrics _disk_io;
 
 public:
